Fix build breaks.
createwebstackclient needed some care to build in its new home.
In particular: its function name was colliding with the basic
creator in webstackclient.h, and it needed the MUJINWEBSTACKCLIENT_API
attribute.

Mujin definitions is removed -- its not needed for webstack and
planning client provides it.

Mujin JSON too has been superceeded by its planning client counterpart,
if both are included (as they are in the vision manager, through a myriad
of inclusion paths). Some spooky action at a distance: we inherit from
MujinControllerCommonException if there is an inclusion path that includes
controller common's mujinjson.
TODO(document/team#86): fix this (or document it). It's better to define what class
mujinjson::MujinJSONException should inherit from as the macro than just
replacing the class. This can also ensure that mujinjsonwebstack::MujinJSONException
is always a std::exception (which might be better).