Split out a webstack client from the controller client.
This is a direct approach to simply split the repos. There are a
lot of potential cleanups since the majority of this codebase is
actually unused.

After the planning client is similarly prepared, we can delete
some of the functionality that both clients would otherwise
export; then this branch represents just the split.