Recover the TaskResource.
This makes the API a little unintuitive, but not in a way that's
fixed in Python, perhaps. In particular: the distinction between
"planning" and "doing a task" is slightly unclear.

For now, the TaskResource is mostly written as it was, but since
the planning client doesn't inherit from it, we have slight changes.
In particular, asking a TaskResource for a PlanningResultResource
is nonsensical in this world.

(Also this is making the word "resource" feel increasingly meaningless.)